requirements.txt
fastapi
uvicorn
httpx==0.27.2
python-dotenv
pydantic-settings
